Revert "Add get holiday"
