+++ .gitignore
@@ -15,4 +15,3 @@ buildSrc/build
 .cxx
 local.properties
 .idea
-/.idea/
