--- apikey.properties
@@ -1,1 +1,0 @@
-holiday_api_key = "88a2a8cc787c4a5bb28b5367c9753c2d"
