+++ app/src/main/java/jt/projects/perfectday/di/Di.kt
@@ -3,22 +3,21 @@ package jt.projects.perfectday.di
 import android.content.Context
 import androidx.room.Room
 import jt.projects.perfectday.App
-import jt.projects.perfectday.interactors.*
+import jt.projects.perfectday.interactors.BirthdayFromPhoneInteractorImpl
+import jt.projects.perfectday.interactors.GetFriendsFromVkUseCase
+import jt.projects.perfectday.interactors.ScheduledEventInteractorImpl
+import jt.projects.perfectday.interactors.SimpleNoticeInteractorImpl
 import jt.projects.perfectday.presentation.calendar.CalendarViewModel
 import jt.projects.perfectday.presentation.calendar.dateFragment.ChosenDateViewModel
 import jt.projects.perfectday.presentation.schedule_event.ScheduleEventViewModel
 import jt.projects.perfectday.presentation.settings.SettingsViewModel
 import jt.projects.perfectday.presentation.today.TodayViewModel
-import jt.projects.repository.network.retrofit.DataSourceHoliday
-import jt.projects.repository.network.retrofit.facts.FactsRepoImpl
-import jt.projects.repository.network.retrofit.facts.FactsRepository
-import jt.projects.repository.network.retrofit.holiday.HolidayRepository
-import jt.projects.repository.network.retrofit.holiday.HolidayRepositoryImpl
-import jt.projects.repository.network.retrofit.holiday.RetrofitHolidayImpl
-import jt.projects.repository.network.retrofit.holiday.dto.HolidayDTO
+import jt.projects.repository.retrofit.facts.FactsRepoImpl
+import jt.projects.repository.retrofit.facts.FactsRepository
 import jt.projects.repository.room.LocalRepository
 import jt.projects.repository.room.RoomDatabaseImpl
 import jt.projects.repository.room.ScheduledEventDatabase
+
 import jt.projects.utils.network.OnlineStatusLiveData
 import jt.projects.utils.shared_preferences.SimpleSettingsPreferences
 import jt.projects.utils.shared_preferences.SimpleSharedPref
@@ -65,9 +64,6 @@ val roomModule = module {
 val repoModule = module {
     single<FactsRepository> { FactsRepoImpl() }
     single<LocalRepository> { RoomDatabaseImpl(dao = get()) }
-
-    single<DataSourceHoliday<List<HolidayDTO>>> { RetrofitHolidayImpl() }
-    single<HolidayRepository> {HolidayRepositoryImpl(dataSource = get())}
 }
 
 
@@ -76,7 +72,6 @@ val interactorsModule = module {
     single { SimpleNoticeInteractorImpl(repository = get()) }
     single { GetFriendsFromVkUseCase(vkNetworkRepository = get()) }
     single { ScheduledEventInteractorImpl(repository = get()) }
-    single { HolidayInteractorImpl(repository = get()) }
 }
 
 
@@ -95,8 +90,7 @@ val viewModelModule = module {
             birthdayFromPhoneInteractor = get(),
             simpleNoticeInteractorImpl = get(),
             getFriendsFromVkUseCase = get(),
-            scheduledEventInteractorImpl = get(),
-            holidayInteractorImpl = get()
+            scheduledEventInteractorImpl = get()
         )
     }
 
@@ -106,8 +100,7 @@ val viewModelModule = module {
             birthdayFromPhoneInteractor = get(),
             simpleNoticeInteractorImpl = get(),
             getFriendsFromVkUseCase = get(),
-            scheduledEventInteractorImpl = get(),
-            holidayInteractorImpl = get()
+            scheduledEventInteractorImpl = get()
         )
     }
 
@@ -117,8 +110,7 @@ val viewModelModule = module {
             birthdayFromPhoneInteractor = get(),
             simpleNoticeInteractorImpl = get(),
             getFriendsFromVkUseCase = get(),
-            scheduledEventInteractorImpl = get(),
-            holidayInteractorImpl = get()
+            scheduledEventInteractorImpl = get()
         )
     }
 }
