+++ app/src/main/java/jt/projects/perfectday/interactors/SimpleNoticeInteractorImpl.kt
@@ -2,23 +2,21 @@ package jt.projects.perfectday.interactors
 
 
 import jt.projects.model.DataModel
-import jt.projects.repository.network.retrofit.facts.FactsRepository
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
+import jt.projects.repository.retrofit.facts.FactsRepository
 import java.time.LocalDate
 
 class SimpleNoticeInteractorImpl(
     private val repository: FactsRepository
 ) {
 
-    suspend fun getFactsByDate(date: LocalDate, factsCount: Int): Flow<DataModel> {
+    suspend fun getFactsByDate(date: LocalDate, factsCount: Int): List<DataModel.SimpleNotice> {
         val result = mutableListOf<DataModel.SimpleNotice>()
         for (i in 1..factsCount) {
             result.add(repository.getFactByDate(date))
         }
-        return flow { result.distinctBy {
+        return result.distinctBy {
             it.description
-        }}
+        }
     }
 
     fun getFakeFacts(): List<DataModel.SimpleNotice> {
