--- repository/src/main/java/jt/projects/repository/network/retrofit/BaseRetrofit.kt
@@ -1,35 +1,0 @@
-package jt.projects.repository.network.retrofit
-
-import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
-import okhttp3.Interceptor
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-class BaseRetrofit {
-
-    fun getService(interceptor: Interceptor, baseUrl:String): HolidayApi {
-        return createRetrofit(interceptor,baseUrl).create(HolidayApi::class.java)
-    }
-
-    private fun createRetrofit(interceptor: Interceptor, baseUrl:String): Retrofit {
-        return Retrofit.Builder()
-            .baseUrl(baseUrl)
-            .addConverterFactory(GsonConverterFactory.create())
-            .addCallAdapterFactory(CoroutineCallAdapterFactory())
-            .client(createOkHttpClient(interceptor))
-            .build()
-    }
-
-    private fun createOkHttpClient(interceptor: Interceptor): OkHttpClient {
-        val httpClient = OkHttpClient.Builder()
-        httpClient.addInterceptor(interceptor)
-        httpClient.addInterceptor(HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY))
-        return httpClient.build()
-    }
-
-    companion object{
-        val newInstansRetrofit = BaseRetrofit()
-    }
-}
