--- repository/src/main/java/jt/projects/repository/network/retrofit/DataSourceHoliday.kt
@@ -1,7 +1,0 @@
-package jt.projects.repository.network.retrofit
-
-import java.time.LocalDate
-
-interface DataSourceHoliday<T> {
-    suspend fun getData(apiKey:String, country:String, date:LocalDate):T
-}
