--- repository/src/main/java/jt/projects/repository/network/retrofit/holiday/HolidayRepositoryImpl.kt
@@ -1,13 +1,0 @@
-package jt.projects.repository.network.retrofit.holiday
-
-import jt.projects.repository.network.retrofit.DataSourceHoliday
-import jt.projects.repository.network.retrofit.holiday.dto.HolidayDTO
-import java.time.LocalDate
-
-
-class HolidayRepositoryImpl(private val dataSource : DataSourceHoliday<List<HolidayDTO>>):HolidayRepository {
-    override suspend fun getHoliday(apiKey:String, country:String,date: LocalDate): List<HolidayDTO> {
-        return dataSource.getData(apiKey,country,date)
-
-    }
-}
