--- repository/src/main/java/jt/projects/repository/network/retrofit/holiday/dto/HolidayDTO.kt
@@ -1,3 +1,0 @@
-package jt.projects.repository.network.retrofit.holiday.dto
-
-class HolidayDTO : ArrayList<HolidayDTOItem>()
