+++ repository/src/main/java/jt/projects/repository/retrofit/BaseInterceptor.kt
@@ -1,4 +1,4 @@
-package jt.projects.repository.network.retrofit
+package jt.projects.repository.retrofit
 
 import okhttp3.Interceptor
 import java.io.IOException
