+++ repository/src/main/java/jt/projects/repository/retrofit/facts/FactsRepository.kt
@@ -1,4 +1,4 @@
-package jt.projects.repository.network.retrofit.facts
+package jt.projects.repository.retrofit.facts
 
 import jt.projects.model.DataModel
 import kotlinx.coroutines.flow.Flow
