+++ repository/src/main/java/jt/projects/repository/room/LocalRepository.kt
@@ -4,7 +4,7 @@ import jt.projects.model.DataModel
 import kotlinx.coroutines.flow.Flow
 
 interface LocalRepository {
-    fun getAll(): Flow<DataModel.ScheduledEvent>
+    suspend fun getAll(): Flow<DataModel.ScheduledEvent>
 
     suspend fun insert(scheduledEvent: DataModel.ScheduledEvent)
