+++ utils/src/main/java/jt/projects/utils/Const.kt
@@ -4,8 +4,6 @@ package jt.projects.utils
  * URLS
  */
 const val FACTS_BASE_URL_LOCATIONS = "http://numbersapi.com/"
-const val HOLIDAY_BASE_URL_LOCATION = "https://holidays.abstractapi.com/"
-const val HOLIDAY_COUNTRY           = "Ru"
 
 /**
  * COMMON SETTINGS
@@ -39,8 +37,3 @@ const val NETWORK_SERVICE = "NETWORK_SERVICE"
  */
 const val FAKE = "FAKE"
 const val REAL = "REAL"
-
-/*
-* FOR DEBUG LO
- */
-val DEBUG: Boolean = BuildConfig.DEBUG && true
